Исправлены ошибки преобразования типов

+++ directProblemQuadratureMethod/basicFunctions.cpp
@@ -7,12 +7,12 @@ using namespace std;
 
 float J_0(const float x)
 {
-	return static_cast<float>(_j0(x));
+	return static_cast<float>(_j0(static_cast<double>(x)));
 }
 
 float N_0(const float x)
 {
-	return static_cast<float>(_y0(x));
+	return static_cast<float>(_y0(static_cast<double>(x)));
 }
 
 void Lasting(const string & st, clock_t & time)
